Add and move everything to SimpleDataTable
